Add alignment to bioloc
Loc is not yet stable. This is the first library I am making for it and
is therefore quite experimental.

A general strategy for porting new material to Loc is to first determine
the types of operations (i.e. write the type section). Ideally, the
names chosen should be fairly general. Next alias the general names to
language specific names (alias section). Then write the native language
wrappers (source section).

Difficulties arise with programs that are highly polymorphic, that is,
the type is dependent on the parameters.

Types should not be overly specific. For example, I use `Bioseq` rather
than `Fasta`. Fasta is the universal representation of sequence data,
but within a program, a native representation is used that cannot be
described as Fasta. I am still using simple types, with one value
mapping uniquely to one type. Eventually I shall have to develop a more
sophisticated type system.

For manifolds wrapping programs, `*.version` and `*_exists` manifolds
should be defined.