Add handling for Smof functions
Smof is a fasta handling tool I wrote in Python. It is split into many
subcommands. Each subcommand I represented with one function in Loc.
Most of the subcommands are clean `Bioseq->Bioseq` transforms, but a few
are messier, specifically the doubly polymorphic `smof grep` function.

It might be better to replace all of these with the more atomic
functions available in any of the bio* packages (e.g. bioperl or
bioconductor).